Time: 0 ms (100.00%) | Memory: 9.2 MB (20.97%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -24,6 +24,6 @@ public:
         if (digit > INT_MAX || digit < INT_MIN)
         return 0;
 
-        return (int)digit;
+        return digit;
     }
 };
